Move remaining writeback actions into their file
