+++ frontend/src/metabase/dashboard/actions/actions.js
@@ -860,16 +860,3 @@ export const revertToRevision = createThunkAction(
     };
   },
 );
-
-// Writeback
-export const OPEN_ACTION_PARAMETERS_MODAL =
-  "metabase/data-app/OPEN_ACTION_PARAMETERS_MODAL";
-export const openActionParametersModal = createAction(
-  OPEN_ACTION_PARAMETERS_MODAL,
-);
-
-export const CLOSE_ACTION_PARAMETERS_MODAL =
-  "metabase/data-app/CLOSE_ACTION_PARAMETERS_MODAL";
-export const closeActionParametersModal = createAction(
-  CLOSE_ACTION_PARAMETERS_MODAL,
-);
